changelog/template: Move templates to md format
WARNING! DO NOT USE THIS FILES DIRECTLY!

This is intended for Backend usage only and this is NOT a Markdown
file, it's a Handlebars template that will be fed to our Backend
for processing.

This renamed from tmpl to md for consistency reasons and that we
moved this from Go templating to Handlebars.

Signed-off-by: Raphiel Rollerscaperers <rapherion@raphielgang.org>